+++ src/modules/posts/infra/typeorm/entities/IgnoredTopic.ts
@@ -28,7 +28,7 @@ class IgnoredTopic {
   @JoinColumn({ name: 'post_id', referencedColumnName: 'post_id' })
   post: Post;
 
-  @Column({ type: 'integer', array: true, default: [] })
+  @Column({ type: 'bigint', array: true, default: [] })
   ignoring: number[];
 
   @CreateDateColumn()
