+++ src/modules/posts/infra/typeorm/entities/PostHistory.ts
@@ -45,7 +45,7 @@ class PostHistory {
   @Column({ default: false })
   notified: boolean;
 
-  @Column({ type: 'integer', array: true, default: [] })
+  @Column({ type: 'bigint', array: true, default: [] })
   notified_to: number[];
 
   @Column({ default: false })
